+++ doodleDefenseMobile/src/ImmuneFoe.h
@@ -14,7 +14,7 @@
 class ImmuneFoe : public Foe{
 public:
     
-    void typeSetup(int level);
+    void typeSetup(float level);
     //void draw();
     
 };
