+++ doodleDefenseMobile/src/StealthFoe.h
@@ -16,7 +16,7 @@
 class StealthFoe : public Foe{
 public:
     
-    void typeSetup(int level);
+    void typeSetup(float level);
     //void draw();
     void findPath();
